% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ZB18a.R
\name{get_solution}
\alias{get_solution}
\title{Get an Orbital Solution}
\usage{
get_solution(orbital_solution = "ZB18a")
}
\arguments{
\item{orbital_solution}{Character vector with the name of the orbital
solution to use. One of \code{"ZB18a"} (default) from Zeebe and Lourens (2019),
or \code{"La11"} (not yet implemented!).}
}
\description{
Get an Orbital Solution
}
