% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution.R
\name{get_ZB}
\alias{get_ZB}
\title{Get a ZB solution}
\usage{
get_ZB(astronomical_solution = "PT-ZB18a", quiet = FALSE, force = FALSE)
}
\arguments{
\item{astronomical_solution}{Character vector with the name of the desired
solution.}

\item{quiet}{Be quiet?
\itemize{
\item If \code{TRUE}, hide info messages.
\item If \code{FALSE} (the default) print info messages and timing.
}}

\item{force}{Force re-downloading the results, even if the solution is saved
to the cache.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the astronomical solution
input and, in the case of the PT-ZB18a, some preprocessed new columns.
}
\description{
Get a ZB solution
}
\examples{
\donttest{
get_ZBa()
}
}
\seealso{
\code{\link[=prepare_solution]{prepare_solution()}} Processes precession-tilt astronomical
solution input to include helper columns.
}
