% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_solution.R
\name{prepare_solution}
\alias{prepare_solution}
\title{Prepare Orbital Solution}
\usage{
prepare_solution(data)
}
\arguments{
\item{data}{The output of \code{\link[=get_solution]{get_solution()}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the new columns added.
}
\description{
Calculates helper columns from an orbital solution input.
}
\details{
New columns include:
\itemize{
\item lphu Unwrapped longitude of perihelion \eqn{\varpi} (degrees without
jumps).
\item lanu Unwrapped longitude of the ascending node \eqn{\Omega}
(degrees without jumps).
\item hh Variable: \eqn{e\sin(\varpi)}{ee * sin(lph / R2D)}.
\item kk Variable: \eqn{e\cos(\varpi)}{ee * cos(lph / R2D)}.
\item 
\item pp Variable: \eqn{2\sin(0.5I)\sin(\Omega)}{2 * sin(0.5 * inc / R2D) *
* n  R2D)}.
\item 
\item qq Variable: \eqn{2\sin(0.5I)\cos(\Omega)}{2 * sin(0.5 * inc / R2D) *
* n  R2D)}.
\item 
\item cc Helper: \eqn{\cos(I)}{cos(inc / R2D)}.
\item 
\item dd Helper: \eqn{\cos(I)/2}{cos(inc / R2D / 2)}.
\item 
\item nnx, nny, nnz}{The \eqn{x}, \eqn{y}, and \eqn{z}-components of the
Eart's orbit unit normal vector \eqn{\vec{n}}{n}, normal to Earth's
instantaneous orbital plane.
}
}
