% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvec.R
\name{snvec}
\alias{snvec}
\title{Calculate precession and obliquity from an OS, Td, and Ed.}
\usage{
snvec(
  tend = -1000,
  ed = 1,
  td = 0,
  orbital_solution = "ZB18a",
  tres = 0.4,
  tolerance = 1e-07,
  quiet = FALSE
)
}
\arguments{
\item{tend}{The final timestep in -kyr. Defaults to \code{-1000} kyr.}

\item{ed}{Dynamical ellipticity, normalized to modern. Defaults to \code{1.0}.}

\item{td}{Tidal dissipation, normalized to modern. Defaults to \code{0.0}.}

\item{orbital_solution}{Character vector with the name of the orbital
solution to use. One of \code{"ZB18a"} (default) or \code{"La11"} (not yet
implemented!).}

\item{tres}{The output timestep resolution in kyr. Defaults to \code{0.4}.}

\item{tolerance}{The numerical tolerance passed to \code{\link[deSolve:ode]{deSolve::ode()}}'s \code{rtol}
and \code{atol} arguments. Defaults to 1e-7.}

\item{quiet}{Be quiet?
If \code{TRUE}, hide info messages.
If \code{FALSE} (the default) print info messages and timing.}
}
\value{
A \link[tibble:tibble-package]{tibble} with all the computed results (see Details).
}
\description{
Computes climatic precession and obliquity (or tilt) from an orbital
solution (OS) input and input values for dynamical ellipticity (Ed) and
tidal dissipation (Td). It fits a set of ordinary differential equations.
}
\details{
The output is a \link[tibble:tibble-package]{tibble} with the following columns:
\itemize{
\item \code{time} Time in years.
\item \code{sx}, \code{sy}, \code{sz} Input vector s.
\item \code{age} Age in thousands of years ago (ka).
\item \code{nnx}, \code{nny}, \code{nnz} The euler transform of the input vector s.
\item \code{eei}, \code{lphi}, \code{lani} The orbital solution's eccentricity, unwrapped
long periapse, and unwrapped long ascending node, interpolated to
resulting timesteps.
\item \code{tmp} A temporary value that we used for the calculation.
\item \code{epl} The \code{acos(tmp)}.
\item \code{u} The input vector s as a list-column.
\item \code{nv} The vector n as a list-column.
\item \code{up} Vector u', with coordinates relative to phi(t=0) at J2000
\item \code{phi} The true anomaly.
\item \code{cp} Climatic precession.
}
}
\examples{
# default call
snvec()

# a quick one with few timesteps, low resolution, high tolerance
snvec(-1e2, 1, 0, orbital_solution = "ZB18a", tres = 1, tolerance = 1e-4)
}
\references{
Zeebe, R. E. (2022). Reduced Variations in Earth’s and Mars’
Orbital Inclination and Earth’s Obliquity from 58 to 48 Myr ago due to
Solar System Chaos. \emph{The Astronomical Journal}, 164(3), 107.
\url{https://doi.org/10.3847/1538-3881/ac80f8}

Wikipedia page on Orbital Elements: \url{https://en.wikipedia.org/wiki/Orbital_elements}
}
\seealso{
\code{\link[deSolve:ode]{deSolve::ode()}} for the ODE solver that we use.
}
\author{
Ilja J. Kocken and Richard E. Zeebe
}
