% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvecR-package.R
\docType{package}
\name{snvecR-package}
\alias{snvecR}
\alias{snvecR-package}
\title{snvecR: Calculate Earth’s Obliquity and Precession in the Past}
\description{
Easily calculate precession and obliquity from an orbital solution (defaults to ZB18a from Zeebe and Lourens (2019) \doi{10.1126/science.aax0612}) and assumed or reconstructed values for tidal dissipation (Td) and dynamical ellipticity (Ed). This is a translation and adaptation of the C-code in the supplementary material to Zeebe and Lourens (2022) \doi{10.1029/2021PA004349}, with furher details on the methodology described in Zeebe (2022) \doi{10.3847/1538-3881/ac80f8}. The name of the C-routine is snvec, which refers to the key units of computation: spin vector s and orbit normal vector n.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://japhir.github.io/snvecR/}
}

}
\author{
\strong{Maintainer}: Ilja Kocken \email{ikocken@hawaii.edu} (\href{https://orcid.org/0000-0003-2196-8718}{ORCID}) [translator, copyright holder]

Authors:
\itemize{
  \item Richard Zeebe \email{zeebe@soest.hawaii.edu} (\href{https://orcid.org/0000-0003-0806-8387}{ORCID})
}

}
\keyword{internal}
