% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB17}
\alias{ZB17}
\alias{ZB17a}
\alias{ZB17b}
\alias{ZB17c}
\alias{ZB17d}
\alias{ZB17e}
\alias{ZB17f}
\alias{ZB17h}
\alias{ZB17i}
\alias{ZB17j}
\alias{ZB17k}
\alias{ZB17p}
\title{Astronomical Solutions ZB17}
\format{
\subsection{\code{get_solution("ZB17x")}}{

A data frame with 62,501 rows and 3 columns:
\describe{
\item{time}{Time in thousands of years (kyr).}
\item{ecc}{Eccentricity \eqn{e}{e} (unitless).}
\item{inc}{Inclination \eqn{I}{I} (degrees).}
}
}
}
\source{
All astronomical solutions by Zeebe can be found on
\url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html}.

They can be loaded into R quickly, using \code{\link[=get_solution]{get_solution()}}.
}
\description{
The ZB17x eccentricity solutions span the past 100 Myr.
Available solutions include \code{"ZB17a"}, \code{"ZB17b"}, \code{"ZB17c"}, \code{"ZB17d"},
\code{"ZB17e"}, \code{"ZB17f"}, \code{"ZB17h"}, \code{"ZB17i"}, \code{"ZB17j"}, \code{"ZB17k"}, and
\code{"ZB17p"}.
}
\references{
Zeebe, R. E. (2017). Numerical Solutions for the orbital motion of the Solar
System over the Past 100 Myr: Limits and new results. \emph{The Astronomical
Journal}. \doi{10.3847/1538-3881/aa8cce}
}
