% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB18a-300}
\alias{ZB18a-300}
\alias{ZB18a}
\title{Astronomical Solution ZB18a for the Past 300 Myr}
\format{
\subsection{\code{get_solution("ZB18a-300")}}{

A data frame with 187,501 rows and 3 columns:
\describe{
\item{age}{Age in thousands of years before present (ka).}
\item{ecc}{Eccentricity \eqn{e} (unitless).}
\item{inc}{Inclination \eqn{I} (degrees).}
}
}
}
\source{
\itemize{
\item All astronomical solutions by Zeebe can be found on
\url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html}.
They can be loaded into R quickly, using \code{\link[=get_solution]{get_solution()}}.
}
}
\description{
Astronomical Solution ZB18a for the Past 300 Myr
}
\references{
Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
Paleocene–Eocene boundary age constrained by geology and astronomy.
\emph{Science}, 365(6456), 926–929. \doi{10.1126/science.aax0612}.'
}
