% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB18a}
\alias{ZB18a}
\title{Astronomical Solution ZB18a}
\description{
Astronomical Solution ZB18a
}
\references{
Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
Paleocene–Eocene boundary age constrained by geology and astronomy.
\emph{Science}, 365(6456), 926–929. \doi{10.1126/science.aax0612}.'
}
