% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB20}
\alias{ZB20}
\title{Astronomical Solution ZB20}
\description{
Astronomical Solution ZB20
}
\references{
Zeebe, R. E. and Lourens, L. J. (2022). Geologically constrained
astronomical solutions for the Cenozoic era. \emph{Earth and Planetary Science
Letters}. \doi{10.1016/j.epsl.2022.117595}
}
