% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB23}
\alias{ZB23}
\alias{ZB23.Rxx}
\alias{ZB23.R01}
\alias{ZB23.R02}
\alias{ZB23.R03}
\alias{ZB23.R04}
\alias{ZB23.R05}
\alias{ZB23.R06}
\alias{ZB23.R07}
\alias{ZB23.R08}
\alias{ZB23.R09}
\alias{ZB23.R10}
\alias{ZB23.R11}
\alias{ZB23.R12}
\alias{ZB23.R13}
\alias{ZB23.R14}
\alias{ZB23.R15}
\alias{ZB23.R16}
\alias{ZB23.R17}
\alias{ZB23.R18}
\alias{ZB23.R19}
\alias{ZB23.R20}
\alias{ZB23.R21}
\alias{ZB23.R22}
\alias{ZB23.R23}
\alias{ZB23.R24}
\alias{ZB23.R25}
\alias{ZB23.R26}
\alias{ZB23.R27}
\alias{ZB23.R28}
\alias{ZB23.R29}
\alias{ZB23.R30}
\alias{ZB23.R31}
\alias{ZB23.R32}
\alias{ZB23.R33}
\alias{ZB23.R34}
\alias{ZB23.R35}
\alias{ZB23.R36}
\alias{ZB23.R37}
\alias{ZB23.R38}
\alias{ZB23.R39}
\alias{ZB23.R40}
\alias{ZB23.R41}
\alias{ZB23.R42}
\alias{ZB23.R43}
\alias{ZB23.R44}
\alias{ZB23.R45}
\alias{ZB23.R46}
\alias{ZB23.R47}
\alias{ZB23.R48}
\alias{ZB23.R49}
\alias{ZB23.R50}
\alias{ZB23.R51}
\alias{ZB23.R52}
\alias{ZB23.R53}
\alias{ZB23.R54}
\alias{ZB23.R55}
\alias{ZB23.R56}
\alias{ZB23.R57}
\alias{ZB23.R58}
\alias{ZB23.R59}
\alias{ZB23.R60}
\alias{ZB23.R62}
\alias{ZB23.R63}
\alias{ZB23.R64}
\title{Astronomical Solutions ZB23.Rxx}
\format{
\subsection{\code{get_solution("ZB23.Rxx")}}{

A data frame with 8,750,001 rows and 5 columns:
\describe{
\item{time}{Time in thousands of years (kyr).}
\item{ecc}{Eccentricity \eqn{e}{e} (unitless).}
\item{inc}{Inclination \eqn{I}{I} (radians).}
\item{obliquity}{Obliqity \eqn{\epsilon}{epsilon} (radians).}
\item{cp}{Climatic Precession \eqn{e \sin(\bar{\omega})}{e sin(baromega)} (unitless).}
}
}
}
\source{
All astronomical solutions by Zeebe can be found on
\url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html}.

They can be loaded into R quickly, using \code{\link[=get_solution]{get_solution()}}.
}
\description{
The ZB23.Rxx eccentricity solutions spand the past 3.6 Gyr. Available solutions include
\code{"ZB23.R01"} to \code{"ZB23.R60"} and \code{"ZB23.R62"} to \code{"ZB23.R64"}.
}
\references{
Zeebe, R. E. and Lourens, L. J. (2022). Geologically constrained
astronomical solutions for the Cenozoic era. \emph{Earth and Planetary Science
Letters}. \doi{10.1016/j.epsl.2022.117595}
}
