% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution.R
\name{get_solution}
\alias{get_solution}
\title{Get an Astronomical Solution}
\usage{
get_solution(
  astronomical_solution = "full-ZB18a",
  quiet = FALSE,
  force = FALSE
)
}
\arguments{
\item{astronomical_solution}{Character vector with the name of the desired
solution. Defaults to \code{"full-ZB18a"}.}

\item{quiet}{Be quiet?
\itemize{
\item If \code{TRUE}, hide info messages.
\item If \code{FALSE} (the default) print info messages and timing.
}}

\item{force}{Force re-downloading the results, even if the solution is saved
to the cache.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the astronomical solution
(and some preprocessed new columns).
}
\description{
\code{get_solution()} downloads an astronomical solution and stores it in the user's cache
directory. The next time the function is called, it will load the data from
the cache rather than downloading it again.
}
\details{
All astronomical solutions by Zeebe can be found on
\url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html}.
Supported solutions include:
\itemize{
\item \link{full-ZB18a}
\item \link{ZB17}
\item \link{ZB18a_100} and \link{ZB18a_300}
\item \link{ZB20}
\item \link{ZB23}
\item \link{PT-ZB18a}
}

See their respective documentation pages for details and citations.

This also provides a wrapper for \code{\link[astrochron:getLaskar]{astrochron::getLaskar()}} if one of
\code{"La04"}, \code{"La10a"}, \code{"La10b"}, \code{"La10c"}, \code{"La10d"}, or \code{"La11"} is
specified, but converts the output to a \link[tibble:tibble-package]{tibble}.
Note that we do not cache these solutions locally, however.

It is possible to change the location of the cache directory with
\code{options(snvecR.cachedir = "/path/to/cache")}.
}
\examples{
\dontshow{
# set the cachedir to a temporary directory
pth <- withr::local_tempdir(pattern = "snvecR")
withr::local_options(snvecR.cachedir = pth)
}
\donttest{
get_solution("full-ZB18a")    # input for snvec
get_solution("ZB18a-300")     # eccentricity
get_solution("ZB20a")
get_solution("La11")
get_solution("PT-ZB18a(1,1)") # pre-computed precession-tilt
get_solution("ZB23.R01")      # one of the 3.6 Gyr solutions
}
}
\references{
Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
Paleocene–Eocene boundary age constrained by geology and astronomy.
\emph{Science}, 365(6456), 926–929. \doi{10.1126/science.aax0612}.

Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for
paleo-applications utilizing obliquity and precession cycles: The role of
dynamical ellipticity and tidal dissipation. \emph{Paleoceanography and
Paleoclimatology}. \doi{10.1029/2021PA004349}
}
\seealso{
\link{prepare_solution}, \link{full_ZB18a}, \link{ZB17}, \link{ZB18a_100}, \link{ZB18a_300} \link{ZB20}, \link{ZB23}, \link{PT_ZB18a}
}
