% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution.R
\name{get_solution}
\alias{get_solution}
\title{Get an Astronomical Solution}
\usage{
get_solution(astronomical_solution = "PT-ZB18a", quiet = FALSE, force = FALSE)
}
\arguments{
\item{astronomical_solution}{Character vector with the name of the
astronomical solution to use or a \code{data.frame} with a custom astronomical
solution. One of \code{"PT-ZB18a"} (default) from Zeebe and Lourens (2019),
or ZB17\link{a-p} from Zeebe 2017, or ZB20\link{a-d} from Zeebe and Lourens (2022).}

\item{quiet}{Be quiet?
\itemize{
\item If \code{TRUE}, hide info messages.
\item If \code{FALSE} (the default) print info messages and timing.
}}

\item{force}{Force re-downloading the results, even if the solution is saved
to the cache.}
}
\value{
A \link[tibble:tibble-package]{tibble} with
the astronomical solution (and some preprocessed new columns).
}
\description{
Get an Astronomical Solution
}
\examples{
\donttest{
get_solution()
}
}
\references{
Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for
paleo-applications utilizing obliquity and precession cycles: The role of
dynamical ellipticity and tidal dissipation. \emph{Paleoceanography and
Paleoclimatology}. \doi{10.1029/2021PA004349}
}
\seealso{
\code{\link[=get_ZB]{get_ZB()}}
}
