% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution.R
\name{get_ZB}
\alias{get_ZB}
\title{Get a ZB solution}
\usage{
get_ZB(astronomical_solution = "PT-ZB18a", quiet = FALSE, force = FALSE)
}
\arguments{
\item{astronomical_solution}{Character vector with the name of the desired
solution.}

\item{quiet}{Be quiet?
\itemize{
\item If \code{TRUE}, hide info messages.
\item If \code{FALSE} (the default) print info messages and timing.
}}

\item{force}{Force re-downloading the results, even if the solution is saved
to the cache.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the astronomical solution
input and, in the case of the PT-ZB18a, some preprocessed new columns.
}
\description{
Get a ZB solution
}
\examples{
\donttest{
get_ZBa()
}
}
\references{
Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
Paleocene–Eocene boundary age constrained by geology and astronomy.
\emph{Science}, 365(6456), 926–929. \doi{10.1126/science.aax0612}.

Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for
paleo-applications utilizing obliquity and precession cycles: The role of
dynamical ellipticity and tidal dissipation. \emph{Paleoceanography and
Paleoclimatology}. \doi{10.1029/2021PA004349}
}
\seealso{
\code{\link[=prepare_solution]{prepare_solution()}} Processes precession-tilt astronomical
solution input to include helper columns.
}
