% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvec.R
\name{snvec}
\alias{snvec}
\title{Calculate climatic precession and obliquity from OS, Td, and Ed.}
\usage{
snvec(
  tend = -1000,
  ed = 1,
  td = 0,
  orbital_solution = "ZB18a",
  tres = 0.4,
  tolerance = 1e-07,
  quiet = FALSE
)
}
\arguments{
\item{tend}{The final timestep in -kyr. Defaults to \code{-1000} years.}

\item{ed}{Dynamical ellipticity. Defaults to \code{1.0}.}

\item{td}{Tidal dissipation. Defaults to \code{0.0}.}

\item{orbital_solution}{The orbital solution to use. See details.}

\item{tres}{The output timestep resolution in kyr. Defaults to \code{0.4}.}

\item{tolerance}{The numerical tolerance passed to \code{\link[deSolve:ode]{deSolve::ode()}}'s \code{rtol} and \code{atol}.}

\item{quiet}{If quiet, do not print info messages.}
}
\value{
A \link[tibble:tibble-package]{tibble} with all the computed results.
}
\description{
Computes precession and tilt/obliquity from an astronomical solution input
and parameter values for dynamical ellipticity and tidal dissipation.
It fits a set of ordinary differential equations.
}
\details{
Currently only the "ZB18a" orbital solution is supported.
}
\examples{
# default call
snvec()

# a quick one with few timesteps, low resolution, high tolerance
snvec(-1e2, 1, 0, orbital_solution = "ZB18a", tres = 1, tolerance = 1e-4)

}
\author{
Ilja J. Kocken and Richard E. Zeebe
}
