% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvecR-package.R
\docType{package}
\name{snvecR-package}
\alias{snvecR}
\alias{snvecR-package}
\title{snvecR: snvecR: Calculate Climatic Precession and Obliquity from OS, Td, and Ed}
\description{
This package makes it easy to calculate precession and obliquity from an orbital solution (defaults to RZ18a) and assumed or reconstructed values for tidal dissipation (td) and dynamical ellipticity (ed).
}
\author{
\strong{Maintainer}: Ilja Kocken \email{ikocken@hawaii.edu} (\href{https://orcid.org/0000-0003-2196-8718}{ORCID})

Authors:
\itemize{
  \item Richard Zeebe \email{zeebe@hawaii.edu} (\href{https://orcid.org/0000-0003-0806-8387}{ORCID})
}

}
\keyword{internal}
