% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB17}
\alias{ZB17}
\title{Astronomical Solutions ZB17}
\description{
Astronomical Solutions ZB17
}
\references{
Zeebe, R. E. (2017). Numerical Solutions for the orbital motion of the Solar
System over the Past 100 Myr: Limits and new results. \emph{The Astronomical
Journal}. \doi{10.3847/1538-3881/aa8cce}
}
