% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZB18a}
\alias{ZB18a}
\title{Orbital Solution from Zeebe et al., 2018.}
\format{
\subsection{\code{ZB18a}}{

A data frame with 250,001 rows and 8 columns:
\describe{
\item{t}{Time in days.}
\item{age}{Age in thousands of years (kyr) before present.}
\item{aa}{Semimajor axis.}
\item{ee}{Eccentricity.}
\item{inc}{Inclination.}
\item{lph}{Long periapse.}
\item{lan}{Long ascending node.}
\item{arp}{Argument of periapse.}
\item{mna}{Mean anomaly.}
\item{hh}{helper hh.}
\item{kk}{helper hk.}
\item{pp}{helper pp.}
\item{qq}{helper qq.}
\item{cc}{helper cc.}
\item{dd}{helper dd.}
\item{nnx, nny, nnz}{Vector of Earth's orbit normal.}
}
}
}
\source{
All orbital solutions can be found on \url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html}.
The specific one we use here is available at \url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/OS/ZB18a/ems-plan3.dat}.
}
\usage{
ZB18a
}
\description{
The HNBody output of Zeebe & Lourens 2019.
}
\details{
The wikipedia page on \href{https://en.wikipedia.org/wiki/Orbital_elements}{Orbital elements}
describes what the components relate to in order to uniquely specify an orbital plane.
}
\references{
Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
Paleocene–Eocene boundary age constrained by geology and astronomy.
\emph{Science}, 365(6456), 926–929. \url{https://doi.org/10.1126/science.aax0612}
}
\keyword{datasets}
