% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB18a}
\alias{ZB18a}
\alias{get_ZB18a}
\title{Astronomical Solution ZB18a}
\description{
Astronomical Solution ZB18a
}
\references{
Zeebe, R. E. and Lourens, L. J. (2022). A deep-time dating tool for
paleo-applications utilizing obliquity and precession cycles: The role of
dynamical ellipticity and tidal dissipation. \emph{Paleoceanography and
Paleoclimatology}. \doi{10.1029/2021PA004349}
}
