% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZB18a}
\alias{ZB18a}
\title{Orbital Solution ZB18a}
\format{
\subsection{\code{ZB18a}}{

A data frame with 250,001 rows and 20 columns:
\describe{
\item{t}{Time in days.}
\item{age}{Age in thousands of years (kyr) before present.}
\item{aa}{Semimajor axis in astronomical units (au).}
\item{ee}{Eccentricity (unitless).}
\item{inc}{Inclination (degrees).}
\item{lph}{Long periapse (degrees).}
\item{lan}{Long ascending node (degrees).}
\item{arp}{Argument of periapse (degrees).}
\item{mna}{Mean anomaly (degrees).}
The following columns were computed from the above input:
\item{lphu}{Unwrapped long periapse (degrees).}
\item{lanu}{Unwrapped long ascending node (degrees).}
\item{hh}{Helper: \code{ee * sin(lph / R2D)}.}
\item{kk}{Helper: \code{ee * cos(lph / R2D)}.}
\item{pp}{Helper: \code{2 * sin(0.5 * inc / R2D) * sin(lan / R2D)}.}
\item{qq}{Helper: \code{2 * sin(0.5 * inc / R2D) * cos(lan / R2D)}.}
\item{cc}{Helper: \code{cos(inc / R2D)}.}
\item{dd}{Helper: \code{cos(inc / R2D / 2)}.}
\item{nnx, nny, nnz}{Vector of Earth's orbit normal.}
}
}
}
\source{
\itemize{
\item All orbital solutions can be found on \url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html}.
\item The specific one we use here is available at \url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro/PrecTilt/OS/ZB18a/ems-plan3.dat}.
}
}
\usage{
ZB18a
}
\description{
The HNBody output of Zeebe & Lourens (2019).
}
\details{
The wikipedia page on \href{https://en.wikipedia.org/wiki/Orbital_elements}{Orbital elements}
describes what the components relate to in order to uniquely specify an orbital plane.
}
\references{
Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
Paleocene–Eocene boundary age constrained by geology and astronomy.
\emph{Science}, 365(6456), 926–929. \url{https://doi.org/10.1126/science.aax0612}
}
\keyword{datasets}
