% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ZB18a_100}
\alias{ZB18a_100}
\alias{"ZB18a-100"}
\title{Astronomical Solution ZB18a}
\format{
\subsection{\code{get_solution("ZB18a-100")}}{

A data frame with 62,501 rows and 3 columns:
\describe{
\item{time}{Time in thousands of years (kyr).}
\item{ecc}{Eccentricity \eqn{e}{e} (unitless).}
\item{inc}{Inclination \eqn{I}{I} (degrees).}
}
}
}
\source{
All astronomical solutions by Zeebe can be found on
\url{http://www.soest.hawaii.edu/oceanography/faculty/zeebe_files/Astro.html}.

They can be loaded into R quickly, using \code{\link[=get_solution]{get_solution()}}.
}
\description{
The ZB18a_100 eccentricity solution spans the past 100 Myr. See \link{ZB18a_300}
for the past 300 Myr.
}
\references{
Zeebe, R. E., & Lourens, L. J. (2019). Solar System chaos and the
Paleocene–Eocene boundary age constrained by geology and astronomy.
\emph{Science}, 365(6456), 926–929. \doi{10.1126/science.aax0612}.'
}
